- Units are now executed also if no input was specified - True MRT support added: -- UnitInOut is now capable of rendering to all of its outputs. -- The behaviour was tested by creating an only offline unit pipeline with 3d texture as output and specifying the zoffsets explicitely

+++ src/osgPlugins/osgPPU/ReaderWriterPPU.cpp
@@ -67,7 +67,7 @@ private:
                                 {
                                     list.back() = std::pair<osgPPU::Unit*, osg::Texture*>(jt->get(), it->second.get()); 
                                 }else{
-                                    if (inputCount < map.size())
+                                    if ((unsigned int)inputCount < map.size())
                                     {
                                         list.push_back(std::pair<osgPPU::Unit*, osg::Texture*>(jt->get(), it->second.get())); 
                                         inputCount ++;
